+++ src/database/postgresql/migrations/20200811211159_product.js
@@ -2,10 +2,11 @@
 
 exports.up = function (knex) {
   return knex.schema.createTable('products', table => {
-    table.increments('id');
+    table.increments('id').primary();
     table.string('name');
     table.integer('amount');
     table.timestamps(true, true);
+    table.unique('name');
   });
 };