+++ src/database/postgresql/migrations/20200811211202_cart.js
@@ -2,11 +2,12 @@
 
 exports.up = function (knex) {
   return knex.schema.createTable('carts', table => {
-    table.increments('id');
+    table.increments('id').primary();
     table.integer('holder').references('accounts.id');
     table.integer('product').references('products.id');
     table.integer('amount');
     table.timestamps(true, true);
+    table.unique(['holder', 'product']);
   });
 };