+++ tests/Integration/api-rest.test.js
@@ -1,16 +1,17 @@
 'use strict';
 process.env.API_TYPE = 'rest';
-process.env.DATABASE_NAME = 'db_test_restl';
+process.env.DATABASE_NAME = 'db_rest';
 
 const request = require('supertest');
 const randomstring = require('randomstring');
 
-const inMemoryDB = require('../test-utils/inMemoryMongo');
-let db = require('../../src/database');
-db.client = inMemoryDB.mongooseClient;
+const testDB = require('../test-utils/db');
+const db = require('../../src/database');
+db.client = testDB.client || db.client;
 
 const config = require('../../src/config');
 const app = require('../../src/app');
+const { products } = require('../../src/database/mongodb');
 
 function doCall(request) {
   return request
@@ -31,11 +32,11 @@ describe('Integration tests - Rest API', function () {
     requestId = randomstring.generate();
   });
   beforeAll(async () => {
-    await inMemoryDB.connect();
+    await testDB.connect();
   });
   afterAll(async () => {
-    await inMemoryDB.clearDatabase();
-    await inMemoryDB.closeDatabase();
+    await testDB.teardown();
+    await testDB.disconnect(db.client);
     await app.close();
   });
 
@@ -51,7 +52,6 @@ describe('Integration tests - Rest API', function () {
           .expect(200)
           .expect('x-request-id', requestId)
           .expect(res => {
-            console.log(res.headers);
             expect(res.body.data).toHaveProperty('id');
             expect(typeof res.body.data.id).toEqual('string');
             accountId = res.body.data.id;
@@ -79,7 +79,6 @@ describe('Integration tests - Rest API', function () {
           .expect('X-REQUEST-ID', requestId)
           .expect(res => {
             expect(res.body).toHaveProperty('data');
-
             expect(res.body.data).toEqual('Authorization header required');
           });
       });
@@ -103,7 +102,6 @@ describe('Integration tests - Rest API', function () {
         await doCall(request(app).delete('/account/' + temp))
           .auth(`${authUser}`, `${authPwd}`)
           .set('X-REQUEST-ID', requestId)
-
           .expect(200)
           .expect('X-REQUEST-ID', requestId)
           .expect(res => {});
@@ -118,7 +116,6 @@ describe('Integration tests - Rest API', function () {
         await doCall(request(app).post('/product'))
           .auth(`${authUser}`, `${authPwd}`)
           .set('X-REQUEST-ID', requestId)
-
           .send({ name: productA, amount })
           .expect(200)
           .expect('X-REQUEST-ID', requestId)
@@ -132,7 +129,6 @@ describe('Integration tests - Rest API', function () {
         await doCall(request(app).post('/product'))
           .auth(`${authUser}`, `${authPwd}`)
           .set('X-REQUEST-ID', requestId)
-
           .send({ name: productB, amount: 70 })
           .expect(200)
           .expect('X-REQUEST-ID', requestId)
@@ -149,7 +145,6 @@ describe('Integration tests - Rest API', function () {
         await doCall(request(app).patch('/product/' + productId))
           .auth(`${authUser}`, `${authPwd}`)
           .set('X-REQUEST-ID', requestId)
-
           .send({ amount })
           .expect(200)
           .expect('X-REQUEST-ID', requestId)
@@ -163,7 +158,6 @@ describe('Integration tests - Rest API', function () {
       it('responds OK with list of products', async () => {
         await doCall(request(app).get('/products'))
           .set('X-REQUEST-ID', requestId)
-
           .expect(200)
           .expect('X-REQUEST-ID', requestId)
           .expect(res => {
@@ -184,7 +178,6 @@ describe('Integration tests - Rest API', function () {
         const amount = 80;
         await doCall(request(app).post(`/action/hold?accountId=${accountId}&productId=${productId}`))
           .set('X-REQUEST-ID', requestId)
-
           .send({ name: productId, amount })
           .expect(200)
           .expect('X-REQUEST-ID', requestId)
@@ -192,7 +185,6 @@ describe('Integration tests - Rest API', function () {
 
         await doCall(request(app).get('/products'))
           .set('X-REQUEST-ID', requestId)
-
           .expect(200)
           .expect('X-REQUEST-ID', requestId)
           .expect(res => {
@@ -209,13 +201,12 @@ describe('Integration tests - Rest API', function () {
       it('responds 200 with list of account', async () => {
         await doCall(request(app).get(`/account/${accountId}/products`))
           .set('X-REQUEST-ID', requestId)
-
           .expect(200)
           .expect('X-REQUEST-ID', requestId)
           .expect(res => {
             expect(Array.isArray(res.body.data)).toBe(true);
-            const product = res.body.data.find(prod => prod.product === productId);
-            expect(product).toMatchObject({ product: productId, amount: holdAmount });
+            const product = res.body.data.find(prod => prod.product.toString() === productId);
+            expect(product).toMatchObject({ product: productId.toString(), amount: holdAmount });
           });
       });
     });
@@ -225,7 +216,6 @@ describe('Integration tests - Rest API', function () {
         const amount = -35;
         await doCall(request(app).patch(`/action/hold?accountId=${accountId}&productId=${productId}`))
           .set('X-REQUEST-ID', requestId)
-
           .send({ amount })
           .expect(200)
           .expect('X-REQUEST-ID', requestId)
@@ -233,18 +223,16 @@ describe('Integration tests - Rest API', function () {
 
         await doCall(request(app).get(`/account/${accountId}/products`))
           .set('X-REQUEST-ID', requestId)
-
           .expect(200)
           .expect('X-REQUEST-ID', requestId)
           .expect(res => {
             expect(Array.isArray(res.body.data)).toBe(true);
-            const product = res.body.data.find(prod => prod.product === productId);
-            expect(product).toMatchObject({ product: productId, amount: holdAmount + amount });
+            const product = res.body.data.find(prod => prod.product.toString() === productId);
+            expect(product).toMatchObject({ product: productId.toString(), amount: holdAmount + amount });
           });
 
         await doCall(request(app).get('/products'))
           .set('X-REQUEST-ID', requestId)
-
           .expect(200)
           .expect('X-REQUEST-ID', requestId)
           .expect(res => {
@@ -262,14 +250,12 @@ describe('Integration tests - Rest API', function () {
       it('responds with json', async () => {
         await doCall(request(app).delete(`/action/hold?accountId=${accountId}&productId=${productId}`))
           .set('X-REQUEST-ID', requestId)
-
           .expect(200)
           .expect('X-REQUEST-ID', requestId)
           .expect(res => {});
 
         await doCall(request(app).get(`/account/${accountId}/products`))
           .set('X-REQUEST-ID', requestId)
-
           .expect(200)
           .expect('X-REQUEST-ID', requestId)
           .expect(res => {
@@ -278,13 +264,12 @@ describe('Integration tests - Rest API', function () {
 
         await doCall(request(app).get('/products'))
           .set('X-REQUEST-ID', requestId)
-
           .expect(200)
           .expect('X-REQUEST-ID', requestId)
           .expect(res => {
             expect(Array.isArray(res.body.data)).toBe(true);
-            const product = res.body.data.find(prod => prod.id === productId);
-            expect(product).toMatchObject({ id: productId, name: 'product-A', amount: stockAmount });
+            const product = res.body.data.find(prod => prod.id.toString() === productId);
+            expect(product).toMatchObject({ id: productId.toString(), name: 'product-A', amount: stockAmount });
           });
       });
     });
@@ -294,10 +279,10 @@ describe('Integration tests - Rest API', function () {
     // describe('Add new products, list products in stock, hold some products, list again remaining in stock', function () {
     //   it('responds with json', async () => {});
     // });
-    // describe('List products in stock, release some products, list remaining holded, list remaining in stock', function () {
+    // describe('List products in stock, release some products, list remaining held, list remaining in stock', function () {
     //   it('responds with json', async () => {});
     // });
-    // describe('List in stock, List remaining holded, move out all, list remaining holded, list in stock', function () {
+    // describe('List in stock, List remaining held, move out all, list remaining held, list in stock', function () {
     //   it('responds with json', async () => {});
     // });
     // describe('List in stock, hold some products, delete account, list in stock', function () {